backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
bcrypt
python-jose
psycopg2
